Fix imports and renames
