+++ hipfab/__init__.py
@@ -19,4 +19,4 @@
 #  limitations under the License.
 #
 
-from hipchat import *
+from hipfab import *
